Module B/course_work/course_work/requirements.txt
asyncpg
fastapi
ormar
psycopg2-binary
pydantic-settings
uvicorn
starlette
pydantic
databases
SQLAlchemy
jinja2
pyjwt